+++ Magic/CHANGELOG.md
@@ -8,6 +8,7 @@
  - Add wand tagging system for use with requirements, make progressmenu easier to customize
  - Some specific internal FX fixes (please let me know if any effects look weird now)
  - Fix mana/regen boosts applied to classes
+ - /mtrait invisible works on armor stand NPCs
 
 # 7.6.6